250512 10:40 Add feature/book 리뷰 삭제 버튼을 누르면 db에서 삭제되고 일부의 리뷰가 정렬기준으로 보이게하고 더보기를 누르면 나머지 리뷰가 더보기 rows만큼 더보기되는 것을 구현

+++ src/main/java/kr/co/bookhub/vo/Book.java
@@ -22,7 +22,7 @@ public class Book {
 	private String description;
 	private String pubDate;
 	private int reviewCount;
-	private int reviewAvg;
+	private double reviewAvg;
 	private String coverImagePath;
 	private Date createDate;
 	private Date updateDate;
